+++ FluentSecurity.Specification/Helpers/ResetFixture.cs
@@ -1,4 +1,5 @@
 using FluentSecurity.Caching;
+using FluentSecurity.Diagnostics;
 using FluentSecurity.Internals;
 using FluentSecurity.ServiceLocation;
 using NUnit.Framework;
@@ -13,6 +14,7 @@ namespace FluentSecurity.Specification // Do not change the namespace
 		{
 			ServiceLocator.Reset();
 			ExceptionFactory.Reset();
+			SecurityDoctor.Reset();
 			SecurityCache.ClearCache(Lifecycle.HybridHttpContext);
 			SecurityCache.ClearCache(Lifecycle.HybridHttpSession);
 		}