+++ FluentSecurity.Specification/SecurityConfigurationSpec.cs
@@ -1,5 +1,7 @@
 using System;
+using System.Configuration;
 using System.Linq;
+using System.Reflection;
 using FluentSecurity.Specification.Helpers;
 using FluentSecurity.Specification.TestData;
 using NUnit.Framework;
@@ -109,7 +111,7 @@ FluentSecurity.Specification.TestData.BlogController > Index
 		public void Should_throw()
 		{
 			// Arrange
-			var setConfigurationMethod = typeof(SecurityConfiguration).GetMethod("SetConfiguration", System.Reflection.BindingFlags.Static | System.Reflection.BindingFlags.NonPublic);
+			var setConfigurationMethod = typeof(SecurityConfiguration).GetMethod("SetConfiguration", BindingFlags.Static | BindingFlags.NonPublic);
 			var parameters = new System.Collections.Generic.List<object> { null }.ToArray();
 
 			// Act
@@ -117,6 +119,70 @@ FluentSecurity.Specification.TestData.BlogController > Index
 
 			// Assert
 			Assert.That(exception.InnerException, Is.TypeOf(typeof(ArgumentNullException)));
+		}
+	}
+	
+	[TestFixture]
+	[Category("SecurityConfigurationSpec")]
+	public class When_calling_assert_all_actions_on_configuration
+	{
+		[Test]
+		public void Should_throw_exception_when_all_actions_are_not_configured()
+		{
+			//Arrange & Act
+			var securityConfiguration = new SecurityConfiguration(policy => {});
+
+			//Assert
+			Assert.Throws<ConfigurationErrorsException>(() => securityConfiguration.AssertAllActionsAreConfigured());
+		}
+
+		[Test]
+		public void Should_not_throw_exception_when_all_actions_are_configured()
+		{
+			//Arrange & Act
+			var assemblies = (
+				from Assembly assembly in AppDomain.CurrentDomain.GetAssemblies()
+				where !(assembly is System.Reflection.Emit.AssemblyBuilder) &&
+				assembly.GetType().FullName != "System.Reflection.Emit.InternalAssemblyBuilder" &&
+				!assembly.GlobalAssemblyCache 
+				select assembly
+				).ToArray();
+
+			var securityConfiguration = new SecurityConfiguration(policy => policy
+				.ForAllControllersInAssembly(assemblies)
+				.DenyAuthenticatedAccess()
+				);
+				
+			//Assert
+			Assert.DoesNotThrow(() => securityConfiguration.AssertAllActionsAreConfigured());
+		}
+	}
+	
+	[TestFixture]
+	[Category("SecurityConfigurationSpec")]
+	public class When_calling_assert_all_actions_on_configuration_passing_assemblies_as_argument
+	{
+		[Test]
+		public void Should_throw_exception_when_all_actions_are_not_configured()
+		{
+			//Arrange & Act
+			var securityConfiguration = new SecurityConfiguration(policy => {});
+
+			//Assert
+			Assert.Throws<ConfigurationErrorsException>(() => securityConfiguration.AssertAllActionsAreConfigured(new [] { GetType().Assembly }));
+		}
+
+		[Test]
+		public void Should_not_throw_exception_when_all_actions_are_configured()
+		{
+			//Arrange & Act
+			var securityConfiguration = new SecurityConfiguration(policy => policy
+				.ForAllControllers()
+				.DenyAuthenticatedAccess()
+				);
+
+			//Assert
+			Assert.DoesNotThrow(() => securityConfiguration.AssertAllActionsAreConfigured(new [] { GetType().Assembly }));
 		}
 	}
 }