+++ FluentSecurity.Specification/TestData/TestSecurityEventListener.cs
@@ -1,0 +1,10 @@
+﻿using FluentSecurity.Diagnostics;
+using FluentSecurity.Diagnostics.Events;
+
+namespace FluentSecurity.Specification.TestData
+{
+	public class TestSecurityEventListener : ISecurityEventListener
+	{
+		public void Handle(ISecurityEvent securityEvent) {}
+	}
+}