+++ FluentSecurity.TestHelper/Properties/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+﻿using System.Reflection;
+using System.Runtime.CompilerServices;
+[assembly: AssemblyTitle("FluentSecurity.TestHelper")]
+[assembly: InternalsVisibleTo("FluentSecurity.TestHelper.Specification")]