+++ FluentSecurity.TestHelper/Properties/TestHelperAssemblyInfo.cs
@@ -1,2 +1,0 @@
-using System.Runtime.CompilerServices;
-[assembly: InternalsVisibleTo("FluentSecurity.TestHelper.Specification")]