+++ FluentSecurity/Diagnostics/Events/ISecurityEvent.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace FluentSecurity.Diagnostics.Events
+{
+	public interface ISecurityEvent
+	{
+		Guid CorrelationId { get; }
+		string Message { get; }
+		long? CompletedInMilliseconds { get; }
+	}
+}