+++ FluentSecurity/Properties/AssemblyInfo.cs
@@ -1,4 +1,7 @@
+using System.Reflection;
 using System.Runtime.CompilerServices;
+
+[assembly: AssemblyTitle("FluentSecurity")]
 [assembly: InternalsVisibleTo("FluentSecurity.Specification")]
 [assembly: InternalsVisibleTo("FluentSecurity.TestHelper")]
 [assembly: InternalsVisibleTo("FluentSecurity.TestHelper.Specification")]